+++ packages/edit-post/src/index.js
@@ -163,13 +163,5 @@ export { default as __experimentalMainDashboardButton } from './components/heade
 
 export { PluginPostExcerpt };
 
-export function __experimentalPluginPostExcerpt() {
-	deprecated( 'wp.editPost.__experimentalPluginPostExcerpt', {
-		since: '6.6',
-		alternative: 'wp.editPost.PluginPostExcerpt',
-	} );
-	return PluginPostExcerpt;
-}
-
 export { store } from './store';
 export * from './deprecated';
